Add GameManager to allow better level building
The GameManager prefab is supposed to create the most elemental objects
which in turn allows easy testing and building of levels. It is
important that the GameManager is the first Object added to a scene and
that the camera added by default is deleted. After that other prefabs
can be added to the scene.
Introduces the singleton pattern to the GameManager, CameraController,
GUIController and PlayerControllers to get easy access from other
scripts and to prevent duplicates in a scene.